删除了BmobPersistenceServiceImpl的原因在于，从业务角度分析，Bmob提供的支持是数据数据同步单元，而不属于本地持久化层。

+++ src/com/libra/stockanalysisi/engine/serialization/impl/JuheStockInfoSerialization.java
@@ -32,8 +32,8 @@ public class JuheStockInfoSerialization implements IStockInfoSerializtion {
 			Double todayMin = dataObj.getDouble("todayMin");
 			Double competitivePri = dataObj.getDouble("competitivePri");
 			Double reservePri = dataObj.getDouble("reservePri");
-			int traNumber = dataObj.getInt("traNumber");
-			int traAmount = dataObj.getInt("traAmount");
+			Double traNumber = dataObj.getDouble("traNumber");
+			Double traAmount = dataObj.getDouble("traAmount");
 			String strDate = dataObj.getString("date");
 			Date date = new SimpleDateFormat("yyyy-MM-dd").parse(strDate);
 			Stock stock = new Stock(gid, name, todayStartPri, yestodEndPri,
